+++ extensions/heartbeat-core/src/index.ts
@@ -1,6 +1,6 @@
 export { HeartbeatPayload, TodayStats, HeartbeatQueueConfig } from './types';
 export { HeartbeatQueue } from './heartbeat-queue';
-export { postHeartbeats, fetchTodayStats } from './http-client';
+export { postHeartbeats, fetchTodayStats, fetchTodayStatsResult, StatsResult } from './http-client';
 export { getGitBranch } from './git';
 export { getLanguageFromFilePath } from './language-map';
 export { getMachineId } from './machine-id';